+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 
 All notable changes to this project will be documented in this file. This project does not adhere to [Semantic Versioning](http://semver.org/)
+## 2.1.5 (06/22/2016)
+- QA-268: Denied attempt test for Student NSC verifier
+- QA-279: Verify Marketplace Group Affilation page displays correctly
+
 ## 2.1.4 (06/21/2016)
 - QA-277: Add Outline for Verify Group Affiliation page and verify military
 - HOTFIX: Increased wait time for Firefox staging env file